se agrega contador de comentarios ,contador de mensajes, que solo el usuario pueda borrar su mensaje,pero que solo pueda hacerlo si no han pasado mas de 30 min

+++ wall/views.py
@@ -1,5 +1,8 @@
 from django.shortcuts import render , redirect
 from .models import *
+from django.contrib import messages
+from datetime import datetime, time, timedelta
+from django.utils import timezone
 
 def wall(request):
     if 'usuario' not in request.session:
@@ -7,12 +10,8 @@ def wall(request):
     else:
         mensajes = Mensaje.objects.all().order_by('-created_at')
         cantMensajes = len(mensajes)
-        print(cantMensajes)
-        
         comentarios = Comentario.objects.all()
         cantComentarios =len(comentarios)
-
-        print(cantComentarios)
         usuarios = User.objects.all()
         contexto = {
             'mensajes'          : mensajes,
@@ -39,16 +38,23 @@ def crearComentario(request, id):
         comentario = request.POST['comentario'],
         usuario = usuario_comentador,
     )
-    print(request.POST)
-    print(usuario_comentador)
     return redirect('/wall')
 
+def minutos(fecha):
+    today = timezone.now()
+    resultado = (today.year - fecha.year)*365*24*60 + (today.month - fecha.month)*30*24*60 + (today.day - fecha.day)*24*60 + (today.hour-fecha.hour)*60 + (today.minute-fecha.minute)
+    return resultado
+
 def delete(request , id):
-    
+    usuario = User.objects.get(id = request.session['usuario']['id'])
     mensaje = Mensaje.objects.get(id = id)
-    # usuario = User.objects.get(id = request.session['usuario']['id'])
-    # if  mensaje == usuario
-    mensaje.delete()
-    return redirect('/wall')
-    # else:
-    #     return render(request , 'wall.html')
+
+    calcular_tiempo = minutos(mensaje.created_at)
+    if calcular_tiempo > 30:
+        messages.warning(request, "Han pasado mas de 30 min, no puedes borrar el mensaje")
+        return redirect('/wall')
+    else: 
+        messages.success(request ,'Borrado con exito')
+        mensaje.delete()
+        return redirect("/wall")
+
